refactor(imports): break 34 of 46 src import cycles
madge --circular reports 46 circular dependencies in src/ (an earlier graphify
pass undercounted at 19). This removes 34 of them with no behavior change: all
the runtime-relevant barrel-incest cycles plus several type-only ones. The
remaining 12 are benign type-only cycles (store slices, SearchFacet,
CitationExporter) and are deferred to a follow-up.

- Redirect Visualizations container/pane imports off the @/components/Visualizations
  barrel to direct module paths
- Extract NetworkDetails node/link types and the AuthorNetworkGraph/BubblePlot
  hook-component type pairs into leaf *.types.ts files
- Redirect FeedbackForms, Settings/Export, HistogramSlider and useExportFormats
  barrel imports to direct paths
- Drop api/metrics/model's dependency on metrics by typing bibcodes as Bibcode[]
- Move ApiRequestConfig into api/types.ts; api.ts re-exports it
- Extract AffTableState/AffTableAction into AuthorAffiliations/types.ts
- Move UserDataSetterEvent out of the settings/export page into
  components/Settings/Export/types.ts (it was imported by four components)
