+++ src/api/metrics/model.ts
@@ -1,14 +1,12 @@
-import { metricsKeys } from './metrics';
+import { Bibcode } from '@/api/search/types';
 import { IADSApiMetricsParams, IAdsApiMetricsTypes } from './types';
 
-export const getMetricsParams = (bibcodes: Parameters<typeof metricsKeys.primary>[0]): IADSApiMetricsParams => ({
+export const getMetricsParams = (bibcodes: Bibcode[]): IADSApiMetricsParams => ({
   bibcodes: bibcodes,
   types: [IAdsApiMetricsTypes.SIMPLE],
 });
 
-export const getMetricsTimeSeriesParams = (
-  bibcodes: Parameters<typeof metricsKeys.timeSeries>[0],
-): IADSApiMetricsParams => ({
+export const getMetricsTimeSeriesParams = (bibcodes: Bibcode[]): IADSApiMetricsParams => ({
   bibcodes,
   types: [IAdsApiMetricsTypes.INDICATORS, IAdsApiMetricsTypes.TIMESERIES],
 });
