+++ src/components/Visualizations/Containers/OverviewPageContainer.tsx
@@ -18,7 +18,8 @@ import { useRouter } from 'next/router';
 import { ReactElement } from 'react';
 import { FacetField } from '../types';
 import { fqNameYearRange } from '@/components/SearchFacet/YearHistogramSlider';
-import { HIndexGraphPane, YearsGraphPane } from '@/components/Visualizations';
+import { HIndexGraphPane } from '../GraphPanes/HIndexGraphPane';
+import { YearsGraphPane } from '../GraphPanes/YearsGraphPane';
 import { makeSearchParams } from '@/utils/common/search';
 import { IADSApiSearchParams } from '@/api/search/types';
 import { useGetSearchFacet, useGetSearchFacetCounts } from '@/api/search/search';
