+++ src/components/Visualizations/Graphs/AuthorNetworkGraph.types.ts
@@ -1,0 +1,12 @@
+import { HierarchyRectangularNode } from 'd3';
+import { IADSApiAuthorNetworkNode } from '@/api/vis/types';
+
+export interface NetworkHierarchyNode<Datum> extends HierarchyRectangularNode<Datum> {
+  color: string; // cache color data
+}
+
+export interface ILink {
+  source: NetworkHierarchyNode<IADSApiAuthorNetworkNode>;
+  target: NetworkHierarchyNode<IADSApiAuthorNetworkNode>;
+  weight: number;
+}
