+++ src/components/Visualizations/Panes/NetworkDetails/AuthorNetworkDetailsPane.tsx
@@ -8,18 +8,7 @@ import { SummaryPane } from './SummaryPane';
 import { MathJax } from 'better-react-mathjax';
 import { SimpleLink } from '@/components/SimpleLink';
 import { unwrapStringValue } from '@/utils/common/formatters';
-import { IDocsEntity } from '@/api/search/types';
-
-interface Paper extends IDocsEntity {
-  groupAuthorCount?: number;
-}
-
-export interface IAuthorNetworkNodeDetails {
-  type: 'author' | 'group';
-  name: string;
-  papers: Paper[];
-  mostRecentYear: string;
-}
+import { IAuthorNetworkNodeDetails, Paper } from './types';
 
 export type AuthorNetworkDetailsProps = {
   node: IAuthorNetworkNodeDetails;
