+++ src/components/Visualizations/Panes/NetworkDetails/index.ts
@@ -1,2 +1,3 @@
+export * from './types';
 export * from './AuthorNetworkDetailsPane';
 export * from './PaperNetworkDetailsPane';
